requirements.txt
ultralytics
opencv-python
gpiozero
lgpio
python-dotenv
